UPDATE METRICS CALCS TO USE FLOAT DIVISION NOT INTEGER DIVISION
- It was truncating everything down to 0, so changed the calculations
  to use floats.
- Also added a new class, AnalyseAll, to make it easier to re-run all
  the metric calculations.

+++ xyz.struthers.rhul.ham/src/main/java/xyz/struthers/rhul/ham/analysis/AnalyseLgas.java
@@ -212,9 +212,9 @@ public class AnalyseLgas {
 
 		// calculate ratios for each LGA
 		for (String lga : lgaHouseholdCount.keySet()) {
-			housingCostsOver30pc.put(lga, lgaHousingCostsOver30pcCount.get(lga) / lgaHouseholdCount.get(lga));
-			incomeEarnedTop5pc.put(lga, lgaTotalIncomeTop5pc.get(lga) / lgaTotalIncome.get(lga));
-			incomeEarnedTop5pcAU.put(lga, nationalTotalIncomeTop5pc.get(lga) / lgaTotalIncome.get(lga));
+			housingCostsOver30pc.put(lga, Float.valueOf(lgaHousingCostsOver30pcCount.get(lga)) / Float.valueOf(lgaHouseholdCount.get(lga)));
+			incomeEarnedTop5pc.put(lga, Float.valueOf(lgaTotalIncomeTop5pc.get(lga)) / Float.valueOf(lgaTotalIncome.get(lga)));
+			incomeEarnedTop5pcAU.put(lga, Float.valueOf(nationalTotalIncomeTop5pc.get(lga)) / Float.valueOf(lgaTotalIncome.get(lga)));
 		}
 
 		// write LGA metrics to file
